% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_template}
\alias{check_template}
\title{Check template directory to make sure is valid}
\usage{
check_template(repo, local = NULL, working = NULL, clean = TRUE,
  deps = FALSE, test = FALSE)
}
\arguments{
\item{repo}{character with the owner and repo names like: foo/bar}

\item{local}{character with the folder where you want to clone the repo}

\item{working}{folder where the test will be performed. By default one level up}

\item{clean}{boolean whether to remove or not the working directory}

\item{deps}{boolean whether to install the deps of the template}

\item{test}{boolean whether to run the test example}
}
\description{
Check template directory to make sure is valid
}
