% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_template}
\alias{run_template}
\title{Main function to render templates from github or local folders}
\usage{
run_template(repo, local = ".", output_file = NULL, options = list(),
  params_file = NULL, ...)
}
\description{
Main function to render templates from github or local folders
}
